pyproject.toml
[project]
name = "haolib"
version = "0.1.0"
description = "A Python utility library with common components for building applications in the HAO ecosystem"
readme = "README.md"
authors = [{ name = "Kamil Biktimirov" }]
keywords = ["fastapi", "backend", "utilities", "microservices", "observability"]
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "aiohttp>=3.12.15",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "cryptography>=45.0.6",
    "dishka>=1.6.0",
    "fastapi>=0.116.1",
    "fastapi-idempotent>=0.0.3",
    "faststream[confluent]>=0.5.48",
    "httpx>=0.28.1",
    "opentelemetry-distro>=0.57b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.36.0",
    "opentelemetry-exporter-otlp-proto-http>=1.36.0",
    "opentelemetry-instrumentation>=0.57b0",
    "opentelemetry-instrumentation-fastapi>=0.57b0",
    "opentelemetry-instrumentation-faststream>=0.1.3",
    "opentelemetry-instrumentation-httpx>=0.57b0",
    "opentelemetry-instrumentation-logging>=0.57b0",
    "pydantic-settings>=2.10.1",
    "pyjwt[crypto]>=2.10.1",
    "redis>=6.4.0",
    "sentry-sdk>=2.35.0",
    "sqlalchemy>=2.0.42",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
    "uvloop>=0.21.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

# Disable fix for unused imports (`F401`).
lint.unfixable = ["F401"]
lint.select = ["ALL"]
lint.ignore = [
    "D203",    #incorrect-blank-line-before-class
    "D213",    # multi-line-summary-second-line
    "TRY003",  # raise-vanilla-args
    "EM101",   # raw-string-in-exception
    "D107",    # undocumented-public-init
    "A002",    # builtin-argument-shadowing
    "D202",    # blank-line-after-function
    "D401",    # non-imperative-mood
    "G004",    # logging-f-string
    "ANN401",  # any-type
    "S101",    # assert
    "PT011",   # pytest-raises-too-broad
    "B024",    # abstract-base-class-without-abstract-method
    "EXE002",  # shebang-missing-executable-file
    "RUF001",  # ambiguous-unicode-character-string
    "BLE001",  # blind-except
    "ARG001",  # unused-function-argument
    "N818",    # error-suffix-on-exception-name
    "PLR0913", # too-many-arguments
    "TC001",   # typing-only-first-party-import
    "RUF012",  # mutable-class-default
    "COM812",  # missing-trailing-comma
    "FBT001",  # boolean-type-hint-positional-argument
    "FBT002",  # boolean-default-value-positional-argument
]


[tool.coverage.run]
# We need *__dishka_factory* here to be excluded because it is a generated file by dishka.
omit = ["*/lib/*", "*__dishka_factory*", "app.py", "dependencies.py", "tests/*"]


[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.12.9",
]
